requirements.txt
wheel
flask
flask-sqlalchemy
flask-security-too
bcrypt